+++ actionview/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Choices of `select` can optionally contain html attributes as the last element of the child arrays when using grouped/nested collections
+
+        <%= form.select :foo, [["North America", [["United States","US"],["Canada","CA"]], { disabled: "disabled" }]] %>
+        # => <select><optgroup label="North America" disabled="disabled"><option value="US">United States</option><option value="CA">Canada</option></optgroup></select>
+
+    *Chris Gunther*
 
 
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/actionview/CHANGELOG.md) for previous changes.